+++ libs/components/errors/src/lib/modules/error/error-modal-config.ts
@@ -7,15 +7,15 @@ export class ErrorModalConfig {
    * Specifies a title to display in the modal error message.
    * @required
    */
-  public errorTitle: string;
+  public errorTitle: string | undefined;
   /**
    * Specifies a description to provide additional details in the modal error message.
    * @required
    */
-  public errorDescription: string;
+  public errorDescription: string | undefined;
   /**
    * Specifies a label for the action button that closes the modal error message.
    * @required
    */
-  public errorCloseText: string;
+  public errorCloseText: string | undefined;
 }