+++ libs/components/indicators/src/lib/modules/tokens/fixtures/tokens-fixtures.module.ts
@@ -4,11 +4,12 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
 import { SkyTokensModule } from '../tokens.module';
 
+import { SkyTokenTestComponent } from './token.component.fixture';
 import { SkyTokensTestComponent } from './tokens.component.fixture';
 
 @NgModule({
-  declarations: [SkyTokensTestComponent],
+  declarations: [SkyTokenTestComponent, SkyTokensTestComponent],
   imports: [CommonModule, NoopAnimationsModule, SkyTokensModule],
-  exports: [SkyTokensTestComponent],
+  exports: [SkyTokenTestComponent, SkyTokensTestComponent],
 })
 export class SkyTokensFixturesModule {}