+++ libs/components/indicators/testing/src/key-info/key-info-harness-filters.ts
@@ -1,0 +1,7 @@
+import { SkyHarnessFilters } from '@skyux/core/testing';
+
+/**
+ * A set of criteria that can be used to filter a list of SkyKeyInfoHarness instances.
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface SkyKeyInfoHarnessFilters extends SkyHarnessFilters {}