+++ libs/components/pages/src/lib/modules/action-hub/types/action-hub-needs-attention.ts
@@ -1,5 +1,7 @@
 import { NavigationExtras } from '@angular/router';
 
+import { SkyActionHubNeedsAttentionClickHandler } from './action-hub-needs-attention-click-handler';
+
 /**
  * Specifies action items that require attention and directs users to pages
  * where they can resolve them.
@@ -16,11 +18,15 @@ export interface SkyActionHubNeedsAttention {
   /**
    * Specifies a link to resolve the action item.
    */
-  permalink: {
+  permalink?: {
     route?: {
       commands: any[];
       extras?: NavigationExtras;
     };
     url?: string;
   };
+  /**
+   * Specifies a click handler for the action item.
+   */
+  click?: SkyActionHubNeedsAttentionClickHandler;
 }