+++ libs/components/text-editor/src/lib/modules/text-editor/url-modal/text-editor-url-modal.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { SkyModalInstance } from '@skyux/modals';
+import { SkyTabIndex } from '@skyux/tabs';
 import { SkyValidation } from '@skyux/validation';
 
 import { SkyUrlModalContext } from './text-editor-url-modal-context';
@@ -89,8 +90,8 @@ export class SkyTextEditorUrlModalComponent {
     }
   }
 
-  public activeTabChanged(value: number): void {
-    this.activeTab = value;
+  public activeTabChanged(value: SkyTabIndex): void {
+    this.activeTab = Number(value);
   }
 
   public save(): void {