+++ libs/sdk/e2e-schematics/src/generators/component-e2e/index.ts
@@ -10,18 +10,27 @@ import {
   generateFiles,
   joinPathFragments,
   logger,
+  readJson,
   readProjectConfiguration,
   removeDependenciesFromPackageJson,
 } from '@nrwl/devkit';
 import { Linter } from '@nrwl/linter';
+import { moveGenerator } from '@nrwl/workspace';
 import { runTasksInSerial } from '@nrwl/workspace/src/utilities/run-tasks-in-serial';
 
+import { RawWorkspaceJsonConfiguration } from 'nx/src/config/workspace-json-project-json';
+import { updateProjectConfiguration } from 'nx/src/generators/utils/project-configuration';
+import { workspaceFileName } from 'nx/src/project-graph/file-utils';
+
+import { updateJson } from '../../utils';
 import configurePercy from '../configure-percy';
 import configureStorybook from '../configure-storybook';
 import init from '../init';
 import storiesGenerator from '../stories';
 
 import { NormalizedSchema, Schema } from './schema';
+
+const BASE_PATH = 'e2e';
 
 function normalizeOptions(options: Partial<Schema>): NormalizedSchema {
   if (!options.name) {
@@ -46,6 +55,52 @@ function normalizeOptions(options: Partial<Schema>): NormalizedSchema {
 }
 
 /**
+ * Drop the BASE_PATH from the project name and re-sort the projects.
+ */
+function simplifyWorkspaceName(tree: Tree, projectName: string) {
+  const workspaceJson = readJson(tree, workspaceFileName());
+  if (`${BASE_PATH}-${projectName}` in workspaceJson.projects) {
+    const projectConfig = readProjectConfiguration(
+      tree,
+      `${BASE_PATH}-${projectName}`
+    );
+    let projectConfigJson = JSON.stringify(projectConfig).replace(
+      new RegExp(`${BASE_PATH}-${projectName}`, 'g'),
+      projectName
+    );
+    if (projectName.endsWith('-e2e')) {
+      const withoutE2e = projectName.replace(/-e2e$/, '');
+      projectConfigJson = projectConfigJson.replace(
+        new RegExp(`${BASE_PATH}-${withoutE2e}`, 'g'),
+        withoutE2e
+      );
+    }
+    updateProjectConfiguration(
+      tree,
+      `${BASE_PATH}-${projectName}`,
+      JSON.parse(projectConfigJson)
+    );
+
+    updateJson(
+      tree,
+      workspaceFileName(),
+      (json: RawWorkspaceJsonConfiguration) => {
+        json.projects[projectName] =
+          json.projects[`${BASE_PATH}-${projectName}`];
+        delete json.projects[`${BASE_PATH}-${projectName}`];
+        json.projects = Object.keys(json.projects)
+          .sort()
+          .reduce((obj: Record<string, string>, key: string) => {
+            obj[key] = `${json.projects[key]}`;
+            return obj;
+          }, {});
+        return json;
+      }
+    );
+  }
+}
+
+/**
  * - Generates -storybook and -storybook-e2e projects for a component library if they don't already exist.
  * - Applies configuration to the projects.
  * - Generates stories and tests if there are demonstration components in the -storybook project.
@@ -56,6 +111,7 @@ export default async function (tree: Tree, schema: Partial<Schema>) {
   const initTasks = await init(tree, { ansiColor: schema.ansiColor });
 
   let createProject = false;
+  let moveProject = false;
   /* istanbul ignore next */
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   let appGenerator: () => void = () => {};
@@ -63,9 +119,40 @@ export default async function (tree: Tree, schema: Partial<Schema>) {
   let e2eProjectConfig: ProjectConfiguration;
   try {
     projectConfig = readProjectConfiguration(tree, options.storybookAppName);
-    (schema.ansiColor === false ? console.warn : logger.warn)(
-      `The project "${options.storybookAppName}" already exists.`
+    if (projectConfig.root === `apps/${options.storybookAppName}`) {
+      moveProject = true;
+      await moveGenerator(tree, {
+        projectName: options.storybookAppName,
+        destination: `${BASE_PATH}/${options.storybookAppName}`,
+        updateImportPath: false,
+        skipFormat: true,
+      });
+      simplifyWorkspaceName(tree, options.storybookAppName);
+      projectConfig = readProjectConfiguration(tree, options.storybookAppName);
+    }
+    e2eProjectConfig = readProjectConfiguration(
+      tree,
+      `${options.storybookAppName}-e2e`
     );
+    if (e2eProjectConfig.root === `apps/${options.storybookAppName}-e2e`) {
+      moveProject = true;
+      await moveGenerator(tree, {
+        projectName: `${options.storybookAppName}-e2e`,
+        destination: `${BASE_PATH}/${options.storybookAppName}-e2e`,
+        updateImportPath: false,
+        skipFormat: true,
+      });
+      simplifyWorkspaceName(tree, `${options.storybookAppName}-e2e`);
+      e2eProjectConfig = readProjectConfiguration(
+        tree,
+        `${options.storybookAppName}-e2e`
+      );
+    }
+    if (!moveProject) {
+      (schema.ansiColor === false ? console.warn : logger.warn)(
+        `The project "${options.storybookAppName}" already exists.`
+      );
+    }
   } catch (e) {
     createProject = true;
     appGenerator = await applicationGenerator(tree, {
@@ -75,7 +162,18 @@ export default async function (tree: Tree, schema: Partial<Schema>) {
       routing: options.routing,
       strict: options.strict,
       prefix: options.prefix,
+      directory: BASE_PATH,
+      standaloneConfig: true,
     });
+    simplifyWorkspaceName(tree, options.storybookAppName);
+    // The e2e project is generated by the application generator, and gets doubly prefixed.
+    await moveGenerator(tree, {
+      projectName: `${BASE_PATH}-${BASE_PATH}-${options.storybookAppName}-e2e`,
+      destination: `${BASE_PATH}/${options.storybookAppName}-e2e`,
+      updateImportPath: false,
+      skipFormat: true,
+    });
+    simplifyWorkspaceName(tree, `${options.storybookAppName}-e2e`);
     projectConfig = readProjectConfiguration(tree, options.storybookAppName);
     e2eProjectConfig = readProjectConfiguration(
       tree,
@@ -84,6 +182,12 @@ export default async function (tree: Tree, schema: Partial<Schema>) {
 
     // Delete boilerplate files from the storybook project.
     let indexFile = tree.read(`${projectConfig.sourceRoot}/index.html`, 'utf8');
+
+    // istanbul ignore if
+    if (!indexFile) {
+      indexFile = '';
+    }
+
     indexFile = indexFile.replace(
       '<link rel="icon" type="image/x-icon" href="favicon.ico" />',
       ''
@@ -131,15 +235,17 @@ export default async function (tree: Tree, schema: Partial<Schema>) {
       generateCypressSpecs: false,
       linter: Linter.EsLint,
     });
-  } else {
+  } else if (!moveProject) {
     await storiesGenerator(tree, {
       project: options.storybookAppName,
       cypressProject: `${options.storybookAppName}-e2e`,
       generateCypressSpecs: true,
     });
   }
-  await configureStorybook(tree, { name: options.storybookAppName });
-  await configurePercy(tree, { name: `${options.storybookAppName}-e2e` });
+  if (!moveProject) {
+    await configureStorybook(tree, { name: options.storybookAppName });
+    await configurePercy(tree, { name: `${options.storybookAppName}-e2e` });
+  }
 
   // Do not add explicit dependencies for ts-node or webpack.
   const updateDependencies = removeDependenciesFromPackageJson(