+++ libs/sdk/e2e-schematics/src/generators/configure-storybook/index.spec.ts
@@ -2,11 +2,13 @@ import {
   applicationGenerator,
   storybookConfigurationGenerator,
 } from '@nrwl/angular/generators';
-import { getWorkspacePath, readProjectConfiguration } from '@nrwl/devkit';
+import { readProjectConfiguration } from '@nrwl/devkit';
 import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
 import { Linter } from '@nrwl/linter';
 import { TsConfig } from '@nrwl/storybook/src/utils/utilities';
 import { removeGenerator } from '@nrwl/workspace';
+
+import { updateProjectConfiguration } from 'nx/src/generators/utils/project-configuration';
 
 import { updateJson } from '../../utils';
 
@@ -14,7 +16,7 @@ import configureStorybook from './index';
 
 describe('configure-storybook', () => {
   it('should configure storybook', async () => {
-    const tree = createTreeWithEmptyWorkspace(1);
+    const tree = createTreeWithEmptyWorkspace(2);
     tree.write('.gitignore', '#');
     await applicationGenerator(tree, { name: `test-app` });
     await storybookConfigurationGenerator(tree, {
@@ -28,32 +30,37 @@ describe('configure-storybook', () => {
     expect(tree.exists(`apps/test-app/.storybook/preview.js`)).toBeFalsy();
     expect(tree.exists(`apps/test-app/.storybook/preview.ts`)).toBeTruthy();
     const e2eConfig = readProjectConfiguration(tree, `test-app-e2e`);
-    expect(e2eConfig.targets.e2e.options.devServerTarget).toEqual(
+    expect(e2eConfig.targets?.e2e.options.devServerTarget).toEqual(
       `test-app:storybook`
     );
-    expect(e2eConfig.targets.e2e.configurations.ci.skipServe).toBeTruthy();
-    updateJson(tree, getWorkspacePath(tree), (json) => {
-      delete json['projects']['test-app'].architect.build.options;
-      return json;
-    });
+    expect(e2eConfig.targets?.e2e.configurations?.ci.skipServe).toBeTruthy();
+    let testAppConfig = readProjectConfiguration(tree, `test-app`);
+    delete testAppConfig.targets?.build.options;
+    updateProjectConfiguration(tree, `test-app`, testAppConfig);
+    let testE2eAppConfig = readProjectConfiguration(tree, `test-app-e2e`);
+    delete testE2eAppConfig.targets?.e2e.configurations;
+    updateProjectConfiguration(tree, `test-app-e2e`, testE2eAppConfig);
     await configureStorybook(tree, { name: 'test-app' });
     expect(
-      readProjectConfiguration(tree, `test-app`).targets.build.options.styles
+      readProjectConfiguration(tree, `test-app`).targets?.build.options.styles
         .length
     ).toBeGreaterThan(0);
-    updateJson(tree, getWorkspacePath(tree), (json) => {
-      json['projects']['test-app'].architect.build.options = {};
-      return json;
-    });
+    testAppConfig = readProjectConfiguration(tree, `test-app`);
+    testAppConfig.targets = testAppConfig.targets || {};
+    testAppConfig.targets.build.options = {};
+    updateProjectConfiguration(tree, `test-app`, testAppConfig);
+    testE2eAppConfig = readProjectConfiguration(tree, `test-app-e2e`);
+    delete testE2eAppConfig.targets?.e2e.configurations?.ci.skipServe;
+    updateProjectConfiguration(tree, `test-app-e2e`, testE2eAppConfig);
     await configureStorybook(tree, { name: 'test-app' });
     expect(
-      readProjectConfiguration(tree, `test-app`).targets.build.options.styles
+      readProjectConfiguration(tree, `test-app`).targets?.build.options.styles
         .length
     ).toBeGreaterThan(0);
   });
 
   it('should configure storybook tsconfig', async () => {
-    const tree = createTreeWithEmptyWorkspace(1);
+    const tree = createTreeWithEmptyWorkspace(2);
     tree.write('.gitignore', '#');
     await applicationGenerator(tree, { name: `test-app` });
     await storybookConfigurationGenerator(tree, {
@@ -69,7 +76,7 @@ describe('configure-storybook', () => {
   });
 
   it('should configure storybook tsconfig, add include', async () => {
-    const tree = createTreeWithEmptyWorkspace(1);
+    const tree = createTreeWithEmptyWorkspace(2);
     tree.write('.gitignore', '#');
     await applicationGenerator(tree, { name: `test-app` });
     await storybookConfigurationGenerator(tree, {
@@ -90,13 +97,16 @@ describe('configure-storybook', () => {
     await configureStorybook(tree, { name: 'test-app' });
     expect(tree.exists(`apps/test-app/.storybook/tsconfig.json`)).toBeTruthy();
     expect(
-      JSON.parse(tree.read(`apps/test-app/.storybook/tsconfig.json`).toString())
-        .include
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      JSON.parse(
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        tree.read(`apps/test-app/.storybook/tsconfig.json`)!.toString()
+      ).include
     ).toBeTruthy();
   });
 
   it('should skip configuration for non-cypress e2e project', async () => {
-    const tree = createTreeWithEmptyWorkspace(1);
+    const tree = createTreeWithEmptyWorkspace(2);
     tree.write('.gitignore', '#');
     await applicationGenerator(tree, { name: `test-app` });
     await storybookConfigurationGenerator(tree, {
@@ -116,7 +126,8 @@ describe('configure-storybook', () => {
     await configureStorybook(tree, { name: 'test-app' });
     await configureStorybook(tree, { name: 'test-app', ansiColor: false });
     const e2eConfig = readProjectConfiguration(tree, `test-app-e2e`);
-    expect(e2eConfig.targets.e2e).toBeFalsy();
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    expect(e2eConfig.targets!.e2e).toBeFalsy();
     expect(errorSpy).toHaveBeenCalledWith(
       `Project "test-app-e2e" does not have an e2e target with @nrwl/cypress:cypress`
     );