+++ libs/sdk/e2e-schematics/src/generators/stories/index.ts
@@ -49,15 +49,28 @@ function normalizeOptions(
   tree: Tree,
   options: StoriesGeneratorSchema
 ): NormalizedSchema {
+  if (!options.project) {
+    throw new Error('Project name not specified');
+  }
+
   const projects = getProjects(tree);
   const projectConfig = getStorybookProject(tree, options);
-  const projectDirectory = projectConfig.sourceRoot;
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const projectDirectory = projectConfig.sourceRoot!;
   const projectName = options.project;
   const projectRoot = projectConfig.root;
   const e2eProjectConfig = projects.get(
     options.cypressProject || `${projectName}-e2e`
   );
-  const e2eSourceRoot = e2eProjectConfig.sourceRoot;
+
+  let e2eSourceRoot: string | undefined;
+
+  // istanbul ignore else
+  if (e2eProjectConfig) {
+    e2eSourceRoot = e2eProjectConfig.sourceRoot;
+  } else {
+    e2eSourceRoot = undefined;
+  }
 
   return {
     ...options,
@@ -135,9 +148,11 @@ export default async function (tree: Tree, options: StoriesGeneratorSchema) {
 
       // Look for a directory to group this story in
       const paths = filepath
-        .substring(normalizedOptions.projectConfig.sourceRoot.length + 1)
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        .substring(normalizedOptions.projectConfig.sourceRoot!.length + 1)
         .split('/');
-      const filename = paths.pop();
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      const filename = paths.pop()!;
       let componentGroup = '';
       for (let i = paths.length - 1; i >= 0; i--) {
         // Skip directory with the same name as the component or a set of generic names.
@@ -180,7 +195,8 @@ export default async function (tree: Tree, options: StoriesGeneratorSchema) {
         normalizedOptions.projectDirectory,
         componentFilePath
       );
-      if (module && module.module.classDeclaration.name.text) {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      if (module && module.module.classDeclaration.name!.text) {
         let importPath = normalizePath(
           relative(
             filepath.substring(0, filepath.lastIndexOf('/')),
@@ -191,13 +207,15 @@ export default async function (tree: Tree, options: StoriesGeneratorSchema) {
           importPath = `./${importPath}`;
         }
         const moduleImportTransformer = getInsertImportTransformer(
-          module.module.classDeclaration.name.text,
+          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          module.module.classDeclaration.name!.text,
           importPath
         );
         const moduleMetadataImportsTransformer =
           getInsertIdentifierToArrayTransformer(
             'imports',
-            module.module.classDeclaration.name.text
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            module.module.classDeclaration.name!.text
           );
         applyTransformersToPath(tree, filepath, [
           moduleImportTransformer,
@@ -235,11 +253,13 @@ export default async function (tree: Tree, options: StoriesGeneratorSchema) {
           tree.delete(filepath);
           return;
         }
-        let spec = tree.read(filepath, 'utf8');
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        let spec = tree.read(filepath, 'utf8')!;
         const visitIdMatch = spec.match(matchCypressVisitId);
         if (visitIdMatch) {
           const id = visitIdMatch[1];
-          spec = spec
+          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          spec = spec!
             // Inject the theme in the url and use the updated id
             .replace(
               matchCypressVisitId,