+++ libs/sdk/e2e-schematics/src/generators/stories/schema.d.ts
@@ -1,5 +1,5 @@
 export interface StoriesGeneratorSchema {
-  project: string;
+  project?: string;
   cypressProject?: string;
   generateCypressSpecs: boolean;
   paths?: string[];